// src/schemas/Motor.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

/** Motor and inverter data */
export type Motor = {
  /** Motor temperature (°C) */
  motor_temp: number;

  /** Inverter temperature (°C) */
  inverter_temp: number;

  /** IGBT temperature (°C) */
  igbt_temp: number;

  /** Inverter voltage (V) */
  inverters_voltage: number;

  /** Actual motor RPM */
  inverter_actual_rpm: number;

  /** Actual torque from inverter (Nm) */
  inverter_actual_torque: number;

  /** Inverter error code */
  inverter_error_code: number;

  /** Inverter system ready flag */
  inverter_system_ready: boolean;

  /** Inverter DC status flag */
  inverter_dc_on: boolean;

  /** Inverter on flag */
  inverter_on: boolean;

  /** Inverter error flag */
  inverter_error: boolean;

  /** Inverter derating flag */
  inverter_derating: boolean;

  /** Max allowed motor RPM */
  max_rpm: number;

  /** Max allowed torque (Nm) */
  max_torque: number;

  /** Max allowed motor temperature (°C) */
  max_motor_temp: number;

  /** Max allowed inverter temperature (°C) */
  max_inverter_temp: number;

  /** Max allowed IGBT temperature (°C) */
  max_igbp_temp: number;
};